fix: move ErrorButton component to ErrorButton folder
