+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Component } from 'react';
 import SearchSection from './components/SearchSection';
 import CardSection from './components/CardSection/CardSection';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
-import ErrorButton from './components/ErrorButton';
+import ErrorButton from './components/ErrorButton/ErrorButton';
 import Loader from './components/Loader';
 import { fetchCharacters } from './api/api';
 import { type Character } from './ts/interfaces/interfaces';
